requirements.txt
rich
erniebot
pydantic
python-dotenv
openai>=1.6.0
gradio>=4.10.0